# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Parcel tracking service",
        "title": "Delivery Tracker API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Авторизация пользователей",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Авторизация пользователя",
                "parameters": [
                    {
                        "description": "Login request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_dto.LoginUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_dto.LoginUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid login or password",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "User inactive",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/delivery-tracker_internal_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "delivery-tracker_internal_dto.LoginUserRequest": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "delivery-tracker_internal_dto.LoginUserResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
        "delivery-tracker_internal_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}